Fix clippy in cars, renderer, geometry
- needless_range_loop: iterate the fixed-size col_xs/row_ys arrays directly;
  buffer.iter_mut().enumerate() in the raytracer (buffer is built as
  vec![vec![0.0; width]; height], so the index sequence is identical).
- too_many_arguments on render_edges_with_orientation: narrowly-scoped allow.
  It is a pub fn re-exported from raytracer/mod.rs, so folding the params into
  a struct would be a breaking public API change.
- unnecessary_cast in geometry: n_max/n_total are only ever used as f32, so
  declare them f32 and drop the casts. Clippy's own suggestion here is broken
  (it retypes the binding but leaves the integer literal -> E0308) and was not
  applied. 400 and 361 are exactly representable in f32, so bit-identical.
  The separate n_total = 180 (cardioid) is genuinely usize and was left alone.

+++ src/progress/styles/geometry.rs
@@ -377,10 +377,10 @@ impl ProgressStyle for Phyllotaxis {
         let scale = fit_scale(dw, dh);
 
         let golden_angle: f32 = 2.0 * PI * (1.0 - 1.0 / 1.618_033_9); // ≈ 137.508°
-        let n_max: usize = 400;
-        let n_plot = (ctx.eased * n_max as f32).round() as usize;
+        let n_max: f32 = 400.0;
+        let n_plot = (ctx.eased * n_max).round() as usize;
         // c chosen so the outermost seed lands near the grid edge.
-        let c = scale / (n_max as f32).sqrt();
+        let c = scale / n_max.sqrt();
         let rot = ctx.time * 0.15;
 
         for n in 0..n_plot {
@@ -563,8 +563,8 @@ impl ProgressStyle for MaurerRose {
         let k: f32 = 5.0;
         let d_deg: f32 = 71.0; // step in degrees
         let d_rad = d_deg * PI / 180.0;
-        let n_total: usize = 361; // one full revolution in d-degree steps
-        let n_chords = (ctx.eased * n_total as f32).round() as usize;
+        let n_total: f32 = 361.0; // one full revolution in d-degree steps
+        let n_chords = (ctx.eased * n_total).round() as usize;
         let rot = ctx.time * 0.15;
 
         // Compute successive chord endpoints.
